+++ api/src/main/java/com/refurbmarket/controller/CategoryController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.refurbmarket.dto.category.CategoryResponseDto;
+import com.refurbmarket.dto.response.CategoryResponseDto;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
