+++ api/src/main/java/com/refurbmarket/domain/User.java
@@ -10,4 +10,5 @@ public class User {
 	private String name;
 	private String email;
 	private String password;
+	private String phoneNumber;
 }
