+++ api/src/main/java/com/refurbmarket/dto/request/FurnitureSearchByCategoryRequestDto.java
@@ -1,12 +1,12 @@
-package com.refurbmarket.dto.furniture;
+package com.refurbmarket.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public class FurnitureSearchByKeywordRequestDto {
-	private String keyword;
+public class FurnitureSearchByCategoryRequestDto {
+	private Long categoryId;
 	private Integer pageSize = 10;
 	private Integer pageNumber = 0;
 }
