+++ api/src/main/java/com/refurbmarket/dto/request/SignUpRequestDto.java
@@ -1,4 +1,4 @@
-package com.refurbmarket.dto.user;
+package com.refurbmarket.dto.request;
 
 import com.refurbmarket.domain.User;
 
@@ -16,8 +16,10 @@ public class SignUpRequestDto {
 	private String email;
 	@Schema(description = "사용자 비밀번호")
 	private String password;
+	@Schema(description = "사용자 휴대폰 번호")
+	private String phoneNumber;
 
 	public User toDomain() {
-		return new User(null, name, email, password);
+		return new User(null, name, email, password, phoneNumber);
 	}
 }
