+++ api/src/main/java/com/refurbmarket/dto/response/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.refurbmarket.dto.user;
+package com.refurbmarket.dto.response;
 
 import com.refurbmarket.domain.User;
