+++ api/src/test/java/com/refurbmarket/controller/CategoryControllerIntegrationTest.java
@@ -12,7 +12,7 @@ import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
-import com.refurbmarket.dto.category.CategoryResponseDto;
+import com.refurbmarket.dto.response.CategoryResponseDto;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class CategoryControllerIntegrationTest {
