+++ api/src/test/java/com/refurbmarket/controller/FurnitureControllerIntegrationTest.java
@@ -10,10 +10,10 @@ import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
-import com.refurbmarket.dto.furniture.FurnitureListResponseDto;
-import com.refurbmarket.dto.furniture.FurnitureResponseDto;
-import com.refurbmarket.dto.furniture.FurnitureSearchByCategoryRequestDto;
-import com.refurbmarket.dto.furniture.FurnitureSearchByKeywordRequestDto;
+import com.refurbmarket.dto.request.FurnitureSearchByCategoryRequestDto;
+import com.refurbmarket.dto.request.FurnitureSearchByKeywordRequestDto;
+import com.refurbmarket.dto.response.FurnitureListResponseDto;
+import com.refurbmarket.dto.response.FurnitureResponseDto;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class FurnitureControllerIntegrationTest {
